"New text add on 05-25-2022
I added new text to the file abike.txt chronicling what my experience has been so far since yesterday. It has not been a lot of fun, because I did wrong by her and she is trying to process and definitely make decisions. Now, whatever the situation/outcome, we wait and see. Hopefully, she forgives me and it is a positive outcome